Update thorlabs_MCM3000.py
Mostly a tidy up and more thorough testing. Prompted by the previous adaptor sometimes hanging indefinitely if an 'exact' encoder count could not be reached. Adding a +-1 count tolerance to encoder position seems to fix this, since it won't always move to the very last count. Each count is about 200nm so this level of slop should be fine.